new test: this triggers a "no asm.js!' bug
When compiled with asmjs-virtual-asmjs-g++ a.cpp b.cpp

The problem is comdat sections; in particular, the way we don't put our
JavaScript code in the .text* sections but instead build a parallel
universe of .javascript.text* sections; when the .text* section is
comdat, the .javascript.text* section has to be as well, but
apparently (I haven't tested this) gas requires comdat sections to be
declared fully every time you switch to them.

So that means
    .section .javascript%S,"a"
isn't going to cut it, both because it lacks the "G" flag and because it
lacks a comdat identifier.