So I can dlopen() (or something like that) a file and read an int
When you put a shared library on the linker command line as an input
file, it's not treated as an input file; instead, it'll be linked
against as a shared library. I think I tried every command-line option
that sounded even vaguely like it might change that, but none did.

But there's elfedit! It lets you change the type of a live ELF file to
"executable", so ld (which is perfectly happy relinking executables, for
some reason, even though that's not something I think anyone else has
any use for) takes it. Of course that means the dynamic sections will
become corrupted, but we extract the dynamic information anyway and keep
it as JSON-ish JavaScript.

So, roughly, here's how to do it:

$ wasm32-virtual-wasm32-gcc -fvisibility=default -Wl,-shared -fPIC
./tests/156-provide-x.c
$ perl ./wasm-13/ldso.pl ./a.out > dyninfo.js
$ ./wasmrewrite/wasmsect dyninfo.js > dyninfo.wass
$ bash -x ./bin/wasmify-wasm32 a.out > test.wasm
$ cat dyninfo.wass >> test.wasm